requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiohttp==3.9.1
requests==2.31.0
huggingface-hub>=0.23.0,<0.26.0
av==13.1.0
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3

# Transcription backends
faster-whisper==1.1.0
qwen-asr>=0.1.0

# Speaker diarization
pyannote.audio==3.1.1

# Note: Install PyTorch separately with CUDA support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For Qwen3-ASR, optionally install flash-attention for better performance:
# pip install flash-attn --no-build-isolation
